+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+### Added
+ - MoveNextAndExpect<T>() custom assertion.
+ - MoveNextAndExpect(object value) custom assertion.
+
 ## [1.1.0] - 2022-06-16
 ### Added
  - Nested coroutine calls support.
@@ -12,5 +17,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
+[UNRELEASED]: https://github.com/AAulicino/Unity-Coroutines-for-NSubstitute/compare/1.1.0...UNRELEASED
 [1.1.0]: https://github.com/AAulicino/Unity-Coroutines-for-NSubstitute/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/AAulicino/Unity-Coroutines-for-NSubstitute/releases/tag/1.0.0